+++ projects/tflac/tflac.c
@@ -1,0 +1,2 @@
+#define TFLAC_IMPLEMENTATION
+#include "tflac.h"